requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-gigachat==0.3.4
python-dotenv==1.0.1
langgraph==0.2.74